Fix remaining major issues from code review
- Fix protocol version truncation: use u32 for actual version in error
- Add search query validation: empty check and 1000 char limit
- Add publish metadata validation: limits on filename (500), keywords
  (50 max, 100 chars each), and nym_address (500 chars)
- Fix seeder address delimiter: use pipe instead of comma to avoid
  parsing issues if addresses contain commas
- Fix TTL overflow: use subtraction instead of addition in SQL to
  prevent integer overflow when checking expiry

+++ crates/brisby-core/src/error.rs
@@ -17,7 +17,7 @@ pub enum Error {
     Protocol(String),
 
     #[error("Version mismatch: expected {expected}, got {actual}")]
-    VersionMismatch { expected: u8, actual: u8 },
+    VersionMismatch { expected: u8, actual: u32 },
 
     #[error("Decode error: {0}")]
     Decode(#[from] prost::DecodeError),
